Rename storage directory
