+++ content/guides/design-patterns-for-humans.md
