+++ lib/author.js
@@ -1,3 +1,3 @@
-import authors from "storage/authors";
+import authors from "content/authors";
 
 export const findByUsername = (username) => authors.find(author => author.username === username) || {};
