+++ lib/guide.js
@@ -1,6 +1,6 @@
-import guides from 'storage/guides';
-import authors from 'storage/authors';
-import siteConfig from 'storage/site';
+import guides from 'content/guides';
+import authors from 'content/authors';
+import siteConfig from 'content/site';
 
 export const getAllGuides = () => {
   return guides.filter(guide => !guide.draft)
