+++ lib/roadmap.js
@@ -1,4 +1,4 @@
-import roadmaps from "storage/roadmaps";
+import roadmaps from "content/roadmaps";
 
 export const getRequestedRoadmap = req => {
   const normalizedUrl = req.url.replace(/\/$/, '');
