+++ lib/url.js
@@ -1,4 +1,4 @@
-import siteConfig from 'storage/site';
+import siteConfig from 'content/site';
 
 export const prefixHost = (url) => {
   // if it starts with forward slash then prefix site URL
