+++ pages/roadmaps.js
@@ -5,10 +5,10 @@ import DefaultLayout from 'layouts/default/index';
 import SiteNav from 'components/site-nav';
 import PageFooter from 'components/page-footer';
 import PageHeader from 'components/page-header';
-import roadmaps from "storage/roadmaps";
+import roadmaps from "content/roadmaps";
 import RowBlock from 'components/row-block';
 import IconRowBlock from 'components/icon-row-block';
-import siteConfig from "storage/site";
+import siteConfig from "content/site";
 import { BadgeLink, BadgesList, PrimaryBadge, SecondaryBadge } from 'components/badges';
 import Helmet from 'components/helmet';
